feat(clerk-js,shared): Move the retry mechanism inside Session.getToken
This commit moves the exponential retry mechanism inside getToken to make it robust against network hiccups. That way retries will be added for all getToken usages including JWT templates.
