+++ packages/shared/src/utils/errors.ts
@@ -1,0 +1,11 @@
+export function isUnauthorizedError(e: any): boolean {
+  const status = e?.status;
+  const code = e?.errors?.[0]?.code;
+  return code === 'authentication_invalid' && status === 401;
+}
+
+export function isNetworkError(e: any): boolean {
+  // TODO: revise during error handling epic
+  const message = (`${e.message}${e.name}` || '').toLowerCase().replace(/\s+/g, '');
+  return message.includes('networkerror');
+}
