+++ packages/shared/src/utils/index.ts
@@ -20,3 +20,4 @@ export * from './workerTimers';
 export * from './runWithExponentialBackOff';
 export * from './isomorphicAtob';
 export * from './createDeferredPromise';
+export * from './errors';
